+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Core: `build_tool::ensure_steps_are_build` for building a custom set of steps
 * Cli2: `build` (`b`, `make`) subcommand for building a build step
 
+### Removed
+* Cli: `hisho_cli` has been removed.
+
 <!--
 ### Added
 ### Changed