+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 -->
 
+## [1.0.0] - 2023-11-01
+
+Release v1.0
+
 ## [1.0.0-rc.5] - 2023-10-30
 
 ### Added